デグレ修正

+++ RadikoShift/EF/Program.cs
@@ -9,7 +9,7 @@ public partial class Program
     [Key]
     [Column("p_id")]
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-    public int PId { get; set; }
+    public long PId { get; set; }
 
     [Column("id")]
     public string Id { get; set; } = null!;
